move most logic intro transform.js scan use transformation
