+++ old/constructBindMethod.js
